more build related stuff

+++ build/default.ps1
@@ -18,6 +18,17 @@ task Package -depends Test {
 	mkdir ..\targets\packages\lib\net40
 	cp ..\targets\speakeasy\speakeasy.* ..\targets\packages\lib\net40
 
+	# update nuspec package
+	$version = get-content ..\VERSION
+	$when = (get-date).ToString("yyyyMMddHHmmss")
+	$packageVersion = "$version-$buildName-$when"
+
+	get-content ..\targets\packages\package.nuspec | 
+        %{$_ -replace '0.0.0.1', $packageVersion } > ..\targets\packages\package.nuspec.tmp
+	
+	mv ..\targets\packages\package.nuspec.tmp ..\targets\packages\package.nuspec -force
+
+
 	..\src\.nuget\nuget.exe pack "..\targets\packages\package.nuspec" -outputdirectory ".\..\targets\packages"
 }
 
@@ -41,16 +52,19 @@ task Test -depends Clean,CopyTools,Compile {
 task UpdateAssemblyInfo {
 	$version = get-content ..\VERSION
 
+	$assemblyVersion = 'AssemblyVersion("' + $version + '")';
+	$assemblyFileVersion = 'AssemblyFileVersion("' + $version + '")';
+
 	get-content ..\src\CommonAssemblyInfo.cs | 
-        %{$_ -replace 'AssemblyVersion\("[0-9]+(\.([0-9]+|\*)){1,3}"\)', $version } |
-        %{$_ -replace 'AssemblyFileVersion\("[0-9]+(\.([0-9]+|\*)){1,3}"\)', $version }  > ..\src\CommonAssemblyInfo.cs.tmp
+        %{$_ -replace 'AssemblyVersion\("[0-9]+(\.([0-9]+|\*)){1,3}"\)', $assemblyVersion } |
+        %{$_ -replace 'AssemblyFileVersion\("[0-9]+(\.([0-9]+|\*)){1,3}"\)', $assemblyFileVersion }  > ..\src\CommonAssemblyInfo.cs.tmp
 	
 	mv ..\src\CommonAssemblyInfo.cs.tmp ..\src\CommonAssemblyInfo.cs -force
 }
 
 task Compile -depends Clean,UpdateAssemblyInfo { 
   # call msbuild
-  $options = "/p:Configuration=Debug"
+  $options = "/p:Configuration=Release"
   $msbuild = "C:\Windows\Microsoft.NET\Framework64\v4.0.30319\MSBuild.exe"
   
   Push-Location ..\src
@@ -61,5 +75,7 @@ task Compile -depends Clean,UpdateAssemblyInfo {
 }
 
 task Clean { 
-  rm -recurse -force ..\targets
+  if (Test-Path -path '..\targets') {
+	rm -recurse -force ..\targets
+  }
 }
